+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 0.5.2
+ - Plugin options config and cleanups.
+
+## 0.5.1
+ - Options, extended features and stats improvements.
+
 ## 0.5.0
  - Core features prototype.